audit: improve cleanliness, coherence, and modularity
- Fix labor payroll reserve incoherence: distressed firms now keep
  only the operating-week reserve when settling weekly wages,
  matching purchase/maintenance distress logic where policy minimum
  is ignored. Previously payroll retained max(policy, week) even
  for distressed firms, blocking recovery payroll while otherwise
  allowing purchases through the same reserve.

- Standardize fixed-point casts: use cast_unsigned() for Money->u64
  in gameplay candidate variation to satisfy clippy cast_sign_loss,
  and remove useless conversion in harness seed mixing.

- Extract workshop maintenance into dedicated simulation/maintenance.rs
  to reduce 2.9k-line orchestrator (ARCHITECTURE.md updated). This
  isolates condition/quality drift, tool demand, and repair from the
  causal daily pipeline, matching the existing market+purchases split
  and enabling focused maintenance tests.

- Clean simulation pipeline header duplication and clarify public-work
  tool demand documentation: stock gap already transmits price pressure,
  so demand_today write would be dead state after weekly settlement.

No save-schema or public API changes; standard lane green.

+++ src/gameplay/harness.rs
@@ -1072,7 +1072,7 @@ pub(crate) fn jittered_decision_interval(
     sample = sample.wrapping_add(state.clock.day().cast_unsigned());
     sample = sample.wrapping_add(u64::from(accumulator.decision_cycles));
     sample = sample.wrapping_add(u64::from(config.seed_count));
-    sample = sample.wrapping_add(u64::from(config.start_seed & 0xFFFF_FFFF));
+    sample = sample.wrapping_add(config.start_seed & 0xFFFF_FFFF);
     sample = sample.wrapping_add(u64::from(state.player_dynasty_id.value()));
     sample ^= u64::from(accumulator.total_viable_choices).wrapping_mul(0x9E37_79B9_7F4A_7C15);
     sample ^= u64::from(accumulator.quiet_cycles).wrapping_mul(0xBF58_476D_1CE4_E5B9);
